feat: support pro builtin skill debug preparation
