+++ packages/global/core/ai/skill/runtime/builtin.ts
@@ -1,0 +1,9 @@
+export type BuiltinSkillSourceFile = {
+  relativePath: string;
+  content: Buffer;
+};
+
+export type BuiltinSkillSource = {
+  name: string;
+  files: BuiltinSkillSourceFile[];
+};
