+++ packages/service/core/ai/skill/debugChat/index.ts
@@ -1,0 +1,2 @@
+export { handleSkillDebugChat } from './handler';
+export { buildDebugRuntimeNodes } from './runtime';
