+++ packages/service/core/ai/skill/runtime/index.ts
@@ -1,4 +1,4 @@
 export type { DeployedSkillInfo, DeployedSkillVersion } from './types';
 export { getAgentSkillInfos, injectAgentSkillFilesToSandbox } from './core';
-export { getBuiltinSkillsRootPath, injectEditDebugBuiltinSkillsToSandbox } from './builtin';
+export { getBuiltinSkillsRootPath, syncBuiltinSkillsToSandbox } from './builtin';
 export { runAgentSkillVersionEntrypoints } from './entrypoint';
