+++ packages/service/core/workflow/dispatch/ai/agent/index.ts
@@ -33,7 +33,11 @@ import {
 import { i18nT } from '@fastgpt/global/common/i18n/utils';
 import { getErrText } from '@fastgpt/global/common/error/utils';
 import type { InteractiveNodeResponseType } from '@fastgpt/global/core/workflow/template/system/interactive/type';
-import { ensureAgentSandboxRuntime, streamAgentSandboxInitStatus } from './sub/sandbox';
+import {
+  ensureAgentSandboxRuntime,
+  streamAgentSandboxInitStatus,
+  type AgentSandboxPrepareAction
+} from './sub/sandbox';
 import type { WorkflowNodeResponseWriter } from '../../../../chat/nodeResponseStorage';
 import type { RuntimeNodeResponseSummary } from '../../type';
 import { createAgentNodeResponseCollector } from './nodeResponseCollector';
@@ -74,6 +78,7 @@ export type DispatchAgentModuleProps = ModuleDispatchProps<{
   [NodeInputKeyEnum.sandboxEntrypoint]?: string;
 }> & {
   nodeResponseWriter?: WorkflowNodeResponseWriter;
+  agentSandboxPrepareActions?: AgentSandboxPrepareAction[];
 };
 
 type Response = DispatchNodeResultType<{
@@ -134,6 +139,7 @@ export const dispatchRunAgent = async (props: DispatchAgentModuleProps): Promise
     runningAppInfo,
     runningUserInfo,
     workflowStreamResponse,
+    agentSandboxPrepareActions,
     usagePush,
     chatId,
     uid,
@@ -222,6 +228,7 @@ export const dispatchRunAgent = async (props: DispatchAgentModuleProps): Promise
       sandboxEntrypoint: effectiveSandboxEntrypoint,
       skillIds,
       editSkillId,
+      prepareActions: agentSandboxPrepareActions,
       currentFiles: userContext.currentFiles
     });
     // 获取请求上下文
