+++ packages/service/core/workflow/dispatch/ai/agent/sub/sandbox/index.ts
@@ -1,3 +1,7 @@
-export { ensureAgentSandboxRuntime } from './prepare';
+export {
+  createBuiltinSkillPrepareAction,
+  ensureAgentSandboxRuntime,
+  type AgentSandboxPrepareAction
+} from './prepare';
 export { dispatchSandboxTool } from './tool';
 export { streamAgentSandboxInitStatus } from './status';
