+++ packages/service/core/workflow/dispatch/ai/agent/sub/sandbox/prepare.ts
@@ -1,10 +1,11 @@
 import type { AgentInputFile } from '../../adapter/userContext';
 import type { DeployedSkillInfo, DeployedSkillVersion } from '../../../../../../ai/skill/runtime';
+import type { BuiltinSkillSource } from '@fastgpt/global/core/ai/skill/runtime/builtin';
 import {
   getAgentSkillInfos,
   getBuiltinSkillsRootPath,
   injectAgentSkillFilesToSandbox,
-  injectEditDebugBuiltinSkillsToSandbox,
+  syncBuiltinSkillsToSandbox,
   runAgentSkillVersionEntrypoints
 } from '../../../../../../ai/skill/runtime';
 import { prepareAgentSandboxRuntime } from '../../../../../../ai/sandbox/runtime';
@@ -19,6 +20,19 @@ import {
 } from '../../../../../../ai/sandbox/runtime/entrypoint';
 import { resolveSandboxHome } from '../../../../../../ai/sandbox/runtime/home';
 
+export type AgentSandboxPrepareContext = {
+  sandboxClient: SandboxClient;
+  workDirectory: string;
+  currentWorkingDirectory?: string;
+  deployedSkillVersions: DeployedSkillVersion[];
+  skillInfos: DeployedSkillInfo[];
+  skillScanDirectories: string[];
+};
+
+export type AgentSandboxPrepareAction = (
+  context: AgentSandboxPrepareContext
+) => Promise<AgentSandboxPrepareContext>;
+
 type EnsureAgentSandboxRuntimeParams = {
   appId: string;
   userId: string;
@@ -30,18 +44,12 @@ type EnsureAgentSandboxRuntimeParams = {
   sandboxEntrypoint?: string;
   skillIds: string[];
   editSkillId?: string;
+  prepareActions?: AgentSandboxPrepareAction[];
   currentFiles: AgentInputFile[];
 };
 type EnsureAgentSandboxRuntimeResult = {
   sandboxClient?: SandboxClient;
   currentWorkingDirectory?: string;
-  skillInfos: DeployedSkillInfo[];
-};
-type AgentSandboxPrepareContext = {
-  sandboxClient: SandboxClient;
-  workDirectory: string;
-  currentWorkingDirectory?: string;
-  deployedSkillVersions: DeployedSkillVersion[];
   skillInfos: DeployedSkillInfo[];
 };
 type AgentSandboxPrepareStep = (
@@ -64,6 +72,7 @@ export async function ensureAgentSandboxRuntime({
   sandboxEntrypoint,
   skillIds,
   editSkillId,
+  prepareActions = [],
   currentFiles
 }: EnsureAgentSandboxRuntimeParams): Promise<EnsureAgentSandboxRuntimeResult> {
   const sandboxContext = await prepareAgentSandboxRuntime({
@@ -86,14 +95,15 @@ export async function ensureAgentSandboxRuntime({
       const context = {
         ...sandboxContext,
         deployedSkillVersions: [],
-        skillInfos: []
+        skillInfos: [],
+        skillScanDirectories: []
       };
 
       return editSkillId
         ? prepareSandbox(
             context,
             injectCurrentInputFiles(currentFiles),
-            injectEditDebugBuiltinSkills(),
+            ...prepareActions,
             readCurrentWorkingDirectory(),
             scanEditDebugSkillInfos()
           )
@@ -101,6 +111,7 @@ export async function ensureAgentSandboxRuntime({
             context,
             injectSelectedSkillFiles({ teamId, tmbId, skillIds }),
             injectCurrentInputFiles(currentFiles),
+            ...prepareActions,
             readCurrentWorkingDirectory(),
             runSandboxEntrypoint({ sandboxEntrypoint }),
             runSelectedSkillEntrypoints(),
@@ -139,25 +150,53 @@ const readCurrentWorkingDirectory = (): AgentSandboxPrepareStep => async (contex
   currentWorkingDirectory: await readSandboxPwd(context.sandboxClient)
 });
 
-const injectEditDebugBuiltinSkills = (): AgentSandboxPrepareStep => async (context) => {
-  await injectEditDebugBuiltinSkillsToSandbox(context.sandboxClient.provider);
-  return context;
-};
+/**
+ * 创建“同步内置 Skill 到当前 sandbox”的 prepare action。
+ *
+ * 调用方只提供内置 Skill 文件来源；具体同步位置、HOME 解析和后续扫描目录登记
+ * 都在 sandbox prepare 生命周期内完成，避免 API 层感知 sandbox 细节。
+ */
+export const createBuiltinSkillPrepareAction =
+  ({
+    getSources,
+    injectToSandbox = syncBuiltinSkillsToSandbox
+  }: {
+    getSources: () => Promise<BuiltinSkillSource[]>;
+    injectToSandbox?: typeof syncBuiltinSkillsToSandbox;
+  }): AgentSandboxPrepareAction =>
+  async (context) => {
+    const sources = await getSources();
+    if (sources.length === 0) return context;
 
-const scanEditDebugSkillInfos = (): AgentSandboxPrepareStep => async (context) => {
-  const homeDirectory = await resolveSandboxHome(context.sandboxClient.provider);
-  const skillDirectories = homeDirectory
-    ? [context.workDirectory, getBuiltinSkillsRootPath(homeDirectory)]
-    : [context.workDirectory];
+    const homeDirectory = await resolveSandboxHome(context.sandboxClient.provider);
+    if (!homeDirectory) {
+      throw new Error('Failed to resolve sandbox HOME for builtin skill sync');
+    }
 
-  return {
-    ...context,
-    skillInfos: await getAgentSkillInfos({
+    await injectToSandbox({
       sandbox: context.sandboxClient.provider,
-      skillDirectories
-    })
+      homeDirectory,
+      sources
+    });
+
+    const builtinSkillsRootPath = getBuiltinSkillsRootPath(homeDirectory);
+
+    return {
+      ...context,
+      skillScanDirectories: [
+        ...context.skillScanDirectories,
+        ...sources.map((source) => `${builtinSkillsRootPath}/${source.name}`)
+      ]
+    };
   };
-};
+
+const scanEditDebugSkillInfos = (): AgentSandboxPrepareStep => async (context) => ({
+  ...context,
+  skillInfos: await getAgentSkillInfos({
+    sandbox: context.sandboxClient.provider,
+    skillDirectories: [context.workDirectory, ...context.skillScanDirectories]
+  })
+});
 
 const injectSelectedSkillFiles =
   ({
@@ -203,11 +242,16 @@ const runSelectedSkillEntrypoints = (): AgentSandboxPrepareStep => async (contex
 
 const scanSelectedSkillInfos = (): AgentSandboxPrepareStep => async (context) => ({
   ...context,
-  skillInfos:
-    context.deployedSkillVersions.length > 0
-      ? await getAgentSkillInfos({
+  skillInfos: await (() => {
+    const skillDirectories = [
+      ...context.deployedSkillVersions.map(({ targetDir }) => targetDir),
+      ...context.skillScanDirectories
+    ];
+    return skillDirectories.length > 0
+      ? getAgentSkillInfos({
           sandbox: context.sandboxClient.provider,
-          skillDirectories: context.deployedSkillVersions.map(({ targetDir }) => targetDir)
+          skillDirectories
         })
-      : []
+      : Promise.resolve([]);
+  })()
 });
